requirements.txt
Flask>=3.0.0
python-dotenv>=1.0.0
openai>=1.0.0
sentence-transformers>=3.0.0
scikit-learn>=1.3.0
numpy>=1.26.0
gunicorn>=21.2.0

